refacctor view models
